Minor buggfixes on lobby
Two of the same class do no longer spawn when switching character. All names are now displayed when peaple join session.

Name does not update on clients but does on server when switching characters. Player is ready text does not update when switching characters to the correct state. Switching character sets the text to not ready, but the charater is ready