Updates to join match and lobbies list
Testing to set bShouldAdvertise to false and bAllowJoinInProgress to false when host destroys session, to not be able to find them with findsessions. Increment variable from players who have joined a match directly in the onpostlogout check for host to join match instead of clients calling a function in the game mode.