Re-Added Material Logic
MM_TilingPanel is now a master material and has no albedo-texture. It instead uses a packed texture with masks that are colored differently for each part.
Roughness, metallic and AO is now also packed in one texture.

Each instance of the material can change parameters such as metallic, roughness, and the three different parts in the material independently.