More robust and optimized find sessions, join sessions and create sessions + changed singularist skin in lobby
Uses open level to a transition level instead of creating a new session and destroying it imidietly as a transition between sessions. Changed session name to NAME_GameSessions instead of custom name, but keeping custom name as in session settings to be able to find sessions. Made fallback code for if create session fails to forcefully destroy current session and try again to make sure that create session succeds and not leaves player in a ghost session. Updated public lobbies list to find public lobbies when it is set to visible to to get faster results and shortened the timer between updates. Optimized some code in EOSGameInstance, for example made the Keys variables static const global variables instead of private variables. Changed the skin of the singularist in the lobby and changed the animation to gideons animations.